+++ packages/be-ts-lib/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 🔖 [3.0.20] (2026-02-22)
+
+### 🏗️ Changed
+
+- Updated dependencies: now requires `@aneuhold/core-ts-api-lib@^3.0.20`.
+
 ## 🔖 [3.0.19] (2026-02-22)
 
 ### 🏗️ Changed
@@ -211,6 +217,7 @@ No direct code changes; version bump for compatibility with new major versions o
 
 <!-- Link References -->
 
+[3.0.20]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.19...be-ts-lib-v3.0.20
 [3.0.19]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.18...be-ts-lib-v3.0.19
 [3.0.18]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.17...be-ts-lib-v3.0.18
 [3.0.17]: https://github.com/aneuhold/ts-libs/compare/be-ts-lib-v3.0.16...be-ts-lib-v3.0.17